requirements.txt
google-adk
arxiv
google-cloud-aiplatform

